Divided datatypes into (json data structure & UI data structures),
